[CV-423] Removed duplicability by moving CondoUnitAdapter function into PropertyAdapter, and deleting CondoUnitAdapter file
